Language option in CLI
Add -l option to specify which column from "resource/table.tsv".

In order to do that, name lookup has to be moved to output phase as
jp/en translation are usually not update to date, as the result you
get "recipes" with same blank name for different gun/equipment and
the solver does not work with that.

Also need to remove the non-modded version guns if modded one exist
as the id is different.